vt.energy-main1/vt.energy-main/pyproject.toml
[tool.poetry]
name = "vt-api-server"
version = "0.1.0"
description = "API server."
authors = ["Maruthi <maruthi@vidcentum.com>"]
license = "GNU GPL3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sanic = {extras = ["ext"], version = "^23.12.1"}
msgspec = "^0.18.6"
python-frontmatter = "^1.1.0"
pygments = "^2.17.2"
docstring-parser = "^0.16"
libsass = "^0.23.0"
mistune = "^3.0.2"
httpx = "^0.27.0"
PyJWT = {extras = ["crypto"], version = "^2.8.0"}
jinja2 = "^3.1.3"
nats-py = {extras = ["nkeys"], version = "^2.7.2"}
google-cloud-translate = "2.0.1"
streamlit = "^1.33.0"
twilio = "^9.0.5"
pyreadline3 = "^3.4.1"
requests = "^2.32.3"

sanic-ext = "^23.12.0"
cython = "^3.0.10"
slixmpp = "^1.8.5"
winloop = "^0.1.4"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
